Dispensibles - Dead code & lazy guest class
Code smell fix for dead code (removal of route "hello" from __init__.py) and the guest class.  This was my poor design early while learning.  It has become clear though that guest is a feature of auth in our application.
