+++ example/flaskr/templates/base.html
@@ -2,7 +2,7 @@
 <title>{% block title %}{% endblock %} - Ethical Eating</title>
 <link rel="stylesheet" href="{{ url_for('static', filename='css/style.css') }}">
 <nav>
-  <h1><a href="{{ url_for('portal') }}">Ethical Eating</a></h1>
+  <h1><a href="{{ url_for('auth.arrival') }}">Ethical Eating</a></h1>
   <ul>
     {% if g.user %}
       <li><span>{{ g.user['username'] }}</span>
